Moved file GOChoice.h to src/grandorgue/gui/wxcontrols
