+++ src/grandorgue/gui/dialogs/midi-event/GOMidiEventRecvTab.h
@@ -1,6 +1,6 @@
 /*
  * Copyright 2006 Milan Digital Audio LLC
- * Copyright 2009-2024 GrandOrgue contributors (see AUTHORS)
+ * Copyright 2009-2025 GrandOrgue contributors (see AUTHORS)
  * License GPL-2.0 or later
  * (https://www.gnu.org/licenses/old-licenses/gpl-2.0.html).
  */
@@ -13,12 +13,11 @@
 #include <wx/panel.h>
 #include <wx/timer.h>
 
+#include "gui/wxcontrols/GOChoice.h"
 #include "midi/GOMidiCallback.h"
 #include "midi/GOMidiListener.h"
 #include "midi/GOMidiReceiverBase.h"
 #include "modification/GOModificationProxy.h"
-
-#include "GOChoice.h"
 
 class wxButton;
 class wxChoice;
